apps/server/pyproject.toml
[tool.poetry]
name = "devclad-server"
version = "0.1.0"
description = ""
authors = ["Arth Tyagi <arthtyagi7@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
Django = {version = "^4.1", extras = ["argon2"]}
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
psycopg2-binary = "^2.9.3"
channels = "^3.0.5"
boto3 = "^1.24.36"
pytest = "^7.1.2"
pytest-django = "^4.5.2"
djangorestframework-simplejwt = "^5.2.0"
dj-rest-auth = "^2.2.5"
django-allauth = "^0.51.0"
python-decouple = "^3.6"
django-timezone-field = "^5.0"
tzdata = "^2022.2"
Pillow = "^9.2.0"
requests = "^2.28.1"
black = "^22.6.0"
scipy = "^1.9.1"
scikit-learn = "^1.1.2"
django-q = "^1.3.9"
rollbar = "^0.16.3"
croniter = "^1.3.7"
stream-chat = "^4.5.0"
graphene-django = "^3.0.0"
django-graphql-jwt = "^0.3.4"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
